rl_refactored/requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.0.0
numpy>=1.24.0
pygame>=2.5.0
pyyaml>=6.0

